add startup tests (disabled for now)
this is not complete, will test the response when database setup is
bad
